patch: migration for number phone from integer to bigint

+++ db/migrate/20251119214300_add_phone_number_to_users.rb
@@ -1,5 +1,5 @@
 class AddPhoneNumberToUsers < ActiveRecord::Migration[8.1]
   def change
-    add_column :users, :phone_number, :integer unless column_exists?(:users, :phone_number)
+    add_column :users, :phone_number, :bigint unless column_exists?(:users, :phone_number)
   end
 end
